+++ src/libcsg/modules/io/lammpsdatareader.cc
@@ -290,8 +290,8 @@ bool LAMMPSDataReader::MatchFourFieldLabels_(std::vector<std::string> fields,
 
 void LAMMPSDataReader::InitializeAtomAndBeadTypes_() {
   if (!data_.count("Masses")) {
-    string err = "Masses must first be parsed before the atoms can be read.";
-    throw runtime_error(err);
+    throw runtime_error(
+        "Masses must first be parsed before the atoms can be read.");
   }
 
   Index index = 0;
@@ -510,10 +510,9 @@ void LAMMPSDataReader::ReadAtoms_(Topology &top) {
       }
 
       if (atomtypes_.count(atomTypeId) == 0) {
-        string err =
+        throw runtime_error(
             "Unrecognized atomTypeId, the atomtypes map "
-            "may be uninitialized";
-        throw runtime_error(err);
+            "may be uninitialized");
       }
 
       string bead_type_name = atomtypes_[atomTypeId];